Add missing \ in directory path
Because the underscore is a special character in markdown, a `\`
before the underscore is interpreted as escaping the underscore,
and the backslash isn't shown on the page.

By adding two backslashes, the first is still used as an escape
character, this time for the second backslash. The second
backslash is shown on the page.